chore: proper api example
