+++ docs/stacks/api/api.router.ts
@@ -1,6 +1,5 @@
 import type { PackageDocNavigationGroup } from '#navigation/navigation.utils.ts'
-
-import { EXAMPLE_NAVIGATION } from './packages/apiPackages.navigation'
+import { EXAMPLE_NAVIGATION } from '#stacks/api/packages/example/example.navigation.ts'
 
 export const API_PACKAGES_NAVIGATION: PackageDocNavigationGroup[] = [
   {
