+++ docs/stacks/api/packages/example/pages/installation.md
@@ -1,0 +1,1 @@
+### Installation guide here
